pyproject.toml
[tool.poetry]
name = "onboarding agent for ai"
version = "0.1.0"
description = "this agent will call perspective customers and onboard them"
authors = ["Kamal Singh <kamalkantsingh10@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
websocket-client = "^1.8.0"
flask = "^3.1.1"
python-dotenv = "^1.1.0"
notion-client = "^2.3.0"
pydantic = "^2.11.4"
tinydb = "^4.8.2"
pydantic-ai = "^0.2.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
